requirements.txt
discord.py
pyyaml
asyncio
simplejson
pysqlite3
dill
tqdm
gitpython